% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{add_percent_gc}
\alias{add_percent_gc}
\title{Add percent GC content to GRanges metadata}
\usage{
add_percent_gc(gr, genome)
}
\arguments{
\item{gr}{A GRanges object.}

\item{genome}{A BSgenome object.}
}
\value{
GRanges with added percent_gc metadata column.
}
\description{
Add percent GC content to GRanges metadata
}
