% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{checkMaxChrLength}
\alias{checkMaxChrLength}
\title{Cap chromosome lengths to maximum allowed}
\usage{
checkMaxChrLength(len)
}
\arguments{
\item{len}{Numeric vector of chromosome lengths.}
}
\value{
Numeric vector capped at 2^29.
}
\description{
Cap chromosome lengths to maximum allowed
}
