% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{find_peaks}
\alias{find_peaks}
\title{Identify local maxima in a numeric vector}
\usage{
find_peaks(x, m = 3)
}
\arguments{
\item{x}{Numeric vector.}

\item{m}{Integer; window size (default = 3).}
}
\value{
Integer vector of indices corresponding to detected peaks.
}
\description{
Finds peaks (local maxima) using a sliding window.
}
