% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_seqlengths.R
\name{fix_bam_seqlengths}
\alias{fix_bam_seqlengths}
\title{Fix BAM seqlevels and seqlengths using a genome reference}
\usage{
fix_bam_seqlengths(bam_in, bam_out, genome)
}
\arguments{
\item{bam_in}{Character scalar. Path to the input BAM file.}

\item{bam_out}{Character scalar. Path to the output BAM file to be written.}

\item{genome}{An object with valid sequence lengths, typically a
\code{BSgenome} or other object for which \code{GenomeInfoDb::seqlengths()}
is defined.}
}
\value{
Invisibly returns the path to the output BAM file.
}
\description{
This function reads an input BAM file, updates the sequence lengths
in the alignment object to match those in a provided reference genome,
and writes a new BAM file with corrected header information.
}
\examples{
\dontrun{
fix_bam_seqlengths(
  bam_in  = "input.bam",
  bam_out = "output.fixed.bam",
  genome  = BSgenome.Hsapiens.UCSC.hg38
)
}

}
