% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{frag}
\alias{frag}
\title{Fragment size distribution from paired-end BAM files}
\usage{
frag(bamFiles, bamFiles.labels, index = bamFiles)
}
\arguments{
\item{bamFiles}{Character vector of BAM file paths.}

\item{bamFiles.labels}{Character vector of labels matching each BAM file.}

\item{index}{Optional vector of BAM indices (default = same as bamFiles).}
}
\value{
Named list of fragment length distributions.
}
\description{
Computes fragment length distributions from one or more paired-end BAM files.
}
