% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{get_density}
\alias{get_density}
\title{Compute 2D point density for scatterplots}
\usage{
get_density(x, y, ...)
}
\arguments{
\item{x}{Numeric vector of x coordinates.}

\item{y}{Numeric vector of y coordinates.}

\item{...}{Additional arguments passed to \code{MASS::kde2d}.}
}
\value{
Numeric vector of density values corresponding to input points.
}
\description{
Computes local 2D density values for x–y points using kernel density estimation.
}
