% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proseq_metas.R
\name{proseq_3p_meta}
\alias{proseq_3p_meta}
\title{PRO-seq metaplot using only 3' ends of reads}
\usage{
proseq_3p_meta(
  metadata,
  feature_center_gr,
  window = 2000,
  smoothing_span = NULL
)
}
\arguments{
\item{metadata}{data.frame with \code{bam}, \code{sample_name}, \code{condition}.}

\item{feature_center_gr}{GRanges of 1-bp centers to window around.}

\item{window}{Integer bp window half-width (default 2000).}

\item{smoothing_span}{Optional (0,1] loess span for smoothed metaplots.}
}
\description{
PRO-seq metaplot using only 3' ends of reads
}
