% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\docType{data}
\name{t}
\alias{t}
\title{Quick bold ggplot2 theme}
\format{
An object of class \code{theme} (inherits from \code{ggplot2::theme}, \code{gg}, \code{S7_object}) of length 144.
}
\usage{
t
}
\description{
Quick bold ggplot2 theme
}
\keyword{datasets}
