% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{theme_custom}
\alias{theme_custom}
\title{Custom ggplot2 theme for publication-ready plots}
\usage{
theme_custom(
  title_size = 16,
  title_face = "bold",
  subtitle_size = 14,
  subtitle_face = "plain",
  subtitle_color = "red",
  axis_title_size = 12,
  axis_title_face = "bold",
  axis_text_size = 10,
  axis_text_face = "plain",
  legend_text_size = 10,
  legend_text_face = "plain",
  legend_title_size = 12,
  legend_title_face = "bold",
  grid_major = TRUE,
  grid_minor = FALSE
)
}
\description{
Custom ggplot2 theme for publication-ready plots
}
