% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{totalMapped}
\alias{totalMapped}
\title{Compute total mapped reads in a BAM file}
\usage{
totalMapped(bamFile)
}
\arguments{
\item{bamFile}{Path to BAM file.}
}
\value{
Total number of mapped reads.
}
\description{
Compute total mapped reads in a BAM file
}
